When loading web interface assets also handle jar URIs
After assembly we cannot use File objects to access individual asset files. Instead we need to use a JarFileSystem to access the file metadata.

We could also cache the etag and modification dates in case of jars, but this change currently does not do that. We may want to revisit this in the future.
The patch uses a LoadingCache as a simple "lock" when creating the filesystem (uris will overlap and reference mostly the shaded jar, the cache makes sure only one thread is creating the filesystem).

fix #1924