Truncate field values in the result table
With really big field values the browser gets stuck for several minutes
rendering HTML where it looks like it crashed. Now we truncate the fields
in the overview to 2048 characters.

The number is kind of abitrary, but we tested it with that value and big
messages and the browser was behaving okayish. This might be improved in
the future, but right now there isn't much we can do.