Fix actions using custom events from SearchStore (#2067)
We added events to the SearchStore when only part of our code was React,
and everything was disconnected. Now we don't need them any more, and
they are source of some troubles since we made the stores global. In
this commit I'm changing our current code to directly use the
SearchStore.

Fixes #2040