Use a constant layout for dashboards
So far we tried to make the dashboard layout to be responsive with the
screen. This was never really working, as widgets were dancing around
inconsistently. Trying to manage gridster from react made this even
worse, as now it is harder to reset gridster or to make sense when the
DOM is changing.

Instead of trying to unsuccessfully do that, these changes are going in
the opposite direction: Users can unlock the dashboard and arrange
widgets as desired, and that arrangement will be keep in the future,
regardless on the screen size. Users can still scroll through the widgets
or re-arrange them if they want to, but we won't try to be smart about
it.