Display search result table borders in chrome
Starting with Chrome 50, borders are not correctly displayed in the
search result table on the first page rendering. It seems that Chrome
has quite some issues with table and cell borders, specially when
`border-collapse: collapse` is in use.

In this case, I could workaround the issue by changing the table
position to `static`. I didn't notice any difference in Firefox, Safari
or IE.