Include default values in plugable entities forms
Changes in #2114 caused another issue with the pluggable entities forms:
default values were shown in forms, but never submitted to the server.

These changes move the logic taking care of default values to the
`ConfigurationForm` component, also removing incorrect uses of
`defaultValue` in inputs.

In this way, the object containing the data will contain the default
values at the beginning, and they will be overriden by any input the
user type in them.

Refs #2108