Allow URI for X-Graylog-Server-URL to be relative.
This helps us to not rely on having a properly configured, publicly
visible, consistent absolute URL for the Graylog server. This makes
configuration a lot simple for example for setups where both the web and
the REST port are exposed on a single port with a URL prefix for the
latter ("/api" for example), because in this case we can set
X-Graylog-Server-URL to a relative URI, regardless of the hostname.