Normalize paths by default (#2213)
In some situations, the `gl2ServerUrl` configuration option may end in
a slash. As we can't predict when that will happen, this change is
normalizing the path of the generated URI, avoiding URIs with duplicated
slashes in their path, as one of the issues described in #2200.
(cherry picked from commit 70198f279c22a96572807deb31eb56357aea30c2)