Calculate keyword from and to values on the fly
Keyword time ranges may be in memory for a while, and the range they
represent may change over that time. Currently, they are also inconsistent
with the way relative ranges work, which are calculated when `getFrom` and
`getTo` are called.

These changes avoid those two problems by calculating the keyword from and
to values when the methods are called.

Fixes #2301