Remove references to simulation
To avoid tying the main component to something that still lives on a
plugin, we remove the `isSimulation` prop, and all references to
simulation.

Instead, we are adding the `disableMessageActions` prop to make possible
to hide the message actions. We also display some text when the message
index is not set, to let the user know that the message is not
persisted.