Support changes for pipeline processor simulator (#2320)
* Extract message formatting from UniversalSearchStore

Lets other stores receiving messages to format them in the same way.

* Ensure props passed to MessageShow use Immutable.js

As we don't use Immutable.js in new components, convert received props
to not break old components used by MessageShow.

* Allow disabling message preview in message loader

* Adapt MessageDetail to display message simulations

Message simulations will not be persisted, so we remove most of the
actions and links, and adapt some texts.

* Display string representation of field

When displaying a message, use a string representation of fields. This
allows us to render "true" and "false" values, for instance.

* Remove references to simulation

To avoid tying the main component to something that still lives on a
plugin, we remove the `isSimulation` prop, and all references to
simulation.

Instead, we are adding the `disableMessageActions` prop to make possible
to hide the message actions. We also display some text when the message
index is not set, to let the user know that the message is not
persisted.

* Use same object to generate message URL

We generate two URLs to the message inside the `MessageDetail`
component, and we should do it in the same way to avoid inconsistencies.

* Use react-bootstrap components in MessageShow