Do not init available alarm callback types, fetch them explicitly.
When we are fetching the available alarm callback types during init of
AlarmCallbacksStore, we don't have a stream id to fetch for yet.
Therefore we are running into permission problems in the backend
resource for non-admin users. This change replaces the init with an
explicit fetch of available alarm callback types in the consumer.

Fixes #2352