Ensure Reflux uses bluebird to create promises
When using the `triggerPromise` method in a Reflux action, Reflux
creates a new promise underneath using the native promise the browser
provides.

The main reason to avoid this is consistency, as bluebird is our library
of choice for promises, and it's strange seeing a promise that behaves
differently. Most importantly, most browsers' promises only have support
for `then` and `catch` methods, and that makes complicated using other
methods as `finally`, as you need to think if the promise you receive is
a bluebird promise or not.