Do not add Etc area to time zones without area
Adding Etc is incorrect and fails when parsint the zone in the server.
Instead we are adding an unclassified section for organisational purposes,
where we can place all time zones without area.

The area is now removed from those entries in the dropdown and will not
be send to the server.

Fixes #2393