Allow indexing when cluster state is red but write index state green/yellow
The cluster state can turn red even though the current write index is
green. This might happen when running rotation/retention during an ES
maintenance window where not all nodes are present.

For index related tasks we are now checking the state of the current
write index (deflector) instead of the state for all Graylog managed
indices.

The logs and UI are still showing the red cluster state to make sure the
admin will be notified.

Refs #2429
Fixes #2371