Handle situation where more than one index has the deflector alias (#2476)
This can happen when some ES nodes are down and index rotation/retention
kicks in. Once the absent nodes are back, they might bring back an
already deleted index including the deflector alias.

Once we detect more than one index for the deflector alias, we throw an
exception. In that case the index rotation thread tries to cleanup the
aliases by removing it from all indices but the latest one.