Fixing session validation propagation.
When a store is initialising before the `SessionStore` is finished
validating an existing session, it is listening on
`SessionActions.login.completed`. Unfortunately this was not working on
certain cases when there was a race between the promise validating the
session and the initial call to `SessionStore.isLoggedIn()` in
`FetchProvider.json()`. This is fixed by explicitly triggering the
`loginCompleted` action, which is also taking over the state propagation
in the store.