Fixing double analyzer registration.
The PluginStore does not prevent double registration of plugins/exports.
This is a workaround to prevent double registration of field analyzers
due to the package index being included in the core _and_ at least one
plugin. The proper fix would be to prevent double plugin registration,
but this is beyond the scope of this PR.