Fix Version#fromClasspathProperties() when loading from JAR plugin
Loading a `version.properties` file from a JAR plugin didn't work before
because `Resources.getResource()` didn't look into the plugin's class
loader.

This changes the main `#fromClasspathProperties()` method to accept a
Class parameter which will be used to lookup the resource in the correct
class loader.

It also adds `#fromPluginProperties()` a convenience method for plugins.