Change plugin REST resource injection to use Class instances (#2492)
Using the actual resource instances caused injection of all the dependencies,
even when we don't need the actual instance at all the registration simply
needs to have the class. this also prevents early injection of objects which
have a lifecycle but don't manage it (this continues to be a problem,
but cannot be solved on this level).