Pass remote address into HttpHeadersToken (#2556)
The "host" will be the actual client host in case `trusted_proxies` have been set.

For the SSO plugin we need to know the actual remote address, which will be the proxy closest to Graylog, to verify that the proxy is ok to set the user header. Otherwise we cannot verify anything.