Make it possible for plugins to request a shared class loader (#2508)
* Make it possible for plugins to request a shared class loader

In #2280 we changed the plugin loader so that all plugins share one
class loader and they can see each other. (needed for plugin
inter-dependencies)

This is problematic when different plugins have conflicting
dependencies. (see #2436)

With this change, every plugin gets its own class loader by default so
we can avoid dependency clashes. Plugins which depend on other plugins
can request a shared class loader via a graylog-plugin.properties file.

Fixes #2436

* Adjust log messages in PluginLoader

* Move properties file into a subdirectory to avoid clashes

The path to the properties file is stored in the
"Graylog-Plugin-Properties-Path" attribute of the JAR manifest.

* Only create shared class loader if any plugin requests to be shared