Fix widget refresh preference
When the user set the `updateUnfocussed` flag to true, the dashboard
update in the foreground/background button does not reflect the actual
status of the refresh policy, and it also doesn't make any changes in
it.

This commit uses the user preference as default refresh policy, and
let the user temporarily override it by using the update in the
foreground/background button in the dashboard.