Only re-render visualizations when data changes
The widget component now checks if the data from the server is a new
version of the one already in the state, before updating and propagating
the state.

Visualizations also check now if the data passed in their props is
actually new or just the same, avoiding re-rendering visualizations if
it's not necessary.