Make decorator list scrollable
When there are many search decorators in the sidebar it is not possible
to get to the non-visible ones.

This change adapts the sidebar height in a similar way as we do for
field analyzers, so at least it is possible to get to all the decorators
that are applied to the search results.