Enable gzip per default for REST API listener. (#2672)
Due to the consolidation of the web and REST API listeners, this setting
is now becoming much more relevant for the performance of the web
interface. GZIP was disabled per default for the REST API in past
versions, which was a very conservative decision. Now that in certain
configurations the REST API listener also serves the web interface, this
setting is very important for performance, as the web interface assets
are quite big when delivering them uncompressed.

As there is no reason (and no negative experience with GZIP in the
past), it is now enabled per default.

Fixes #2670