Unregister PipelineInterpreter from event bus (#79)
Message decorators and the pipeline simulator create new instances of
PipelineInterpreters that never get garbage collected, as they are still
registered in the event bus.

These changes add a simple workaround for that. We should probably
refactor the lifecycle of the PipelineInterpreter, but this is probably
not the best time to do it.