Fix content pack extractor validation (#2697)
* Correct error message when could not apply content pack

* Do not require target_field in all extractors

Grok and JSON extractors may create several fields, so the
`target_field` will be empty in those extractors.

* Show error message when uploading invalid JSON

When there is an error importing the content pack the user provided,
display an error message with some information about it. It also serves
as a way of getting feedback of the import operation.