General improvements (#2700)
* Rename "Show Original" button to "Show changes"

The button displays the changes applied by the decorators applied in the
stream, so that name seems more logical to me. Also use consistent case
in the button.

* Use larger input for message id field

On the manual message loader, give more width to the message ID
field, so the whole ID is visible.

* Do not use decorators in recent message loader

The recent message loader uses a search to get a recent message, and it
was being decorated.

This commit adds another parameter to `UniversalSearchStore.search()` to
indicate whether or not the search should be decorated.