Rely on Routes to get the replay widget search url
It's better to let `Routes` handle the generation of the replay widget
URL, and this moves the logic to do that.

The change adds the capability to get search or stream search routes
with (optional) query, time range, and resolution parameters. The
incomplete `search_with_query` function is still used in the collector
plugin, so this still needs a bit of cleanup.