Fix 404s on IE 11 using compatibility view (#2782)
When Graylog is running on an intranet, IE 11 may be requesting the site
in compatibility view, depending on the organization and user settings.
That applies even when properly setting the X-UA-Compatible header and
meta tags.

Making requests in compatibility view, IE does not include either
`text/html` or `application/xhtml+xml` MIME types in the `Accept`
header, which are the values we checked in order to deliver the web
interface HTML document when requesting a resource that doesn't exist
(i.e. when clicking on a page number on the search page).

Unfortunately, the way legacy IEs work is also odd in the way it
includes `Accept` headers on requests, being `*/*` the only one that is always
included. You can read more about it here:
https://blogs.msdn.microsoft.com/ieinternals/2009/07/01/ie-and-the-accept-header/

To solve this issue, these changes relax the media types comparison on
`WebAppNotFoundResponseFilter`, by also serving the web interface HTML
document if any of the request headers is compatible with `text/html` or
`application/xhtml+xml`. In that way we also consider that requests with
a wildcard MIME type will be able to handle the HTML document, but we
still avoid serving HTML if the request specified another accepted MIME
type.

Fixes #2768