Improve Grizzly web server setup
- Do not start Grizzly immediately when creating a new HttpServer instance
  to avoid the creation of a dangling thread pool
- Reduce default transport selector runners to 1

We used to start the Grizzly server when calling
GrizzlyHttpServerFactory.createHttpServer(), then modified the
thread pool settings and then started the server again. This resulted
in a dangling thread pool that isn't used at all.

The default transport selector runners is equal to the number of CPU
cores in the system. This might make sense for pure web server work
loads but not for Graylog where we have tons of other work besides serving
the API/web. Change the default to 1 and make the setting configurable.