Simplify pipeline processor UI (#106)
Fixes #104 

* Use pipelines overview page as entry point

* Add connected streams to pipeline overview page

* Move pipeline connection lists to its own component

* Add API method to connect streams to a pipeline

Migrating pipeline connections to the pipeline page also means that
we need to create connections in the reverse order: several streams
connected to a pipeline. I added a new API method for it, as I think the
old method can still be useful for someone using the API
programatically.

* Format PipelineConnectionsActions

* Set connections from pipeline details view

* Use nicer word break in timeline

* Cleanup after moving connections to pipeline view

* Add link to manage rules to rule details page

* Adapt pipeline simulator to recent changes

Now the stream selection needs to be done in the simulate page, as we
don't know before which stream to use.