Removing production flag from build task.
Running webpack with -p/--optimize-minimize for the web interface leads
to mangling of variable names, which breaks rickshaw (as already noted
in ae731ec8e6a95465b8887e5ad86ba031949dcbbc). Therefore the more
specific addition of the UglifyJSPlugin was added to the production
config and -p is not necessary for the webpack run.