Fix NumberField constructor overloading
The previous overloading of NumberField constructors, adding `int` and
`Number` versions of each constructor, fixed the problems with old
plugins not using autoboxing in their constructors, but introduced
ambiguity during compilation.

As for now we only need constructors to take `int` or `double` values,
we overload `NumberField`'s constructor to take those arguments, and
internally we rely on a constructor using `Number`.

Refs #2973