Update field analyzers when search changes
When a new search is executed, fetch new quick values and field
statictics data. In this way the displayed information will be in sync
with the search query, time range, and stream used in the message list
and histogram.

The field graphs were already updating when the search histogram
changed, so they do not require further changes at this point.

Fixes #2908