Improved search indicators
* Make search loading indicator more visible

The indicator is now centered on the top of the screen, and uses the same
colours as a bootstrap info alert.

* Add search retrieval time to search sidebar

Will help to check if the search results were updated when the page
render is fast.

Closes #3031

(cherry picked from commit 8470fc1682744a9c4bb0e9c09c7bf30f2abc473c)