Refresh stream search when stream changes
Most of the time, search results include links to other streams. When
the user is currently searching on a stream, clicking on one of those
links was not working properly: the stream didn't update and the search
was still refering to the old stream.

These changes tackle that problem, by updating the stream when it
changes, and reloading search results as well.

Closes #3033

(cherry picked from commit 2c43b0779958d959f436648f90552f52973fa24a)