Use MongoDB Date to store Alert DateTimes
After the refactoring in 8b15f35f5fef46b0c6b64ec7488e71019708696f, new
Alert documents were using Strings instead of Dates to store
`triggered_at`. That change breaks the sorting of documents by that
field, sorting that we use in some places of the code.

This commit ensures that `triggered_at` and the newly added
`resolved_at` fields are persisted as Date instances, keeping
consistency with the data already stored in the collection.