Honor limit parameter for real
The previous fix was still executing a MongoDB query per stream. The
whole resource was only available for users with permission to read all
streams.

This commit changes that, so it will query MongoDB for alerts in the
streams the user has read permissions. The query is only executed once,
so limit parameter is also working as expected.