Migrate alert receivers (#3120)
* Add list configuration field

We need a field to move email receivers into the email callback
model.

* Add missing prop to DropdownField

* Add support to get list in `Configuration`

* Move email receivers to `EmailAlarmCallback`

- Make user/email receivers part of the `EmailAlarmCallback` model
- Add configuration fields to the callback, ensuring we do not offer
user auto-completion if the user does not have permissions to list users
- Move logic to resolve email addresses to its own class

* Clean up alert receivers

Clean up frontend code related to alert receivers.

* Add Builder to AlarmCallbackConfigurationAVImpl

This allows us to modify the object instead of creating new ones.

* More clean up of alert receivers

This time remove some unused backend code, and adapt previous API to
add/remove receivers to use an `EmailAlarmCallback` if available.

* Convert EmailAlarmCallback migration into a Migration

We need to do this to ensure it runs before the upcoming alert receivers
migration. I also changed the event that gets stored in the cluster
config, as that was still not release and there's no harm running the
migration again in our development machines.

* Move existing alert receivers to EmailAlarmCallback

The migration will move all alert receivers from the streams having at
least a condition and an EmailAlarmCallback.

* Fix condition to run EmailAlarmCallback migrations

Also modified some tests to ensure this works as expected.