Move alert conditions and notifications to alerts page (#3145)
* Make stream more prominent in alerts

* Correct alert condition descriptions

We mixed up the "Field Value Alert Condition" and "Field Content Value
Alert Condition" descriptions #lol

* Change human name for conditions

"Field Content Value Alert Condition" and "Field Value Alert Condition"
are not very descriptive names for those alert conditions, so I am
modifying them to make them a bit more clear. This is only for
presentational purposes, we will keep internal names, at least for now.

* Remove unneded input field

* Add more alert routes

* Move alert conditions to alerts page

* Remove default email alarm callback from test

When testing an alarm callback, only use the alarm callbacks set for the
stream, do not use a default email callback.

* Add new REST resource to handle global callbacks

* Formatting StreamAlertConditionResource

* Add store, actions and routes for new REST resource

* Add feedback to test alert notification action

The test notification action should provide feedback indicating whether
the notification was sent or not.

* Small fixes in alert conditions components

* Add feedback to alarm callback store

* Add alert notifications global page

* Display notifications in condition list

* Remove unused attribute

* Remove alerts link from streams list

This requires a more in-depth clean up, but for now removing the button
and routes is enough.