Add document titles (#3160)
* Update LDAP routes

Use the authentication page to get to LDAP pages. Also delete old (and
unused) LDAP pages.

* Add document titles to core pages

* Simplify DocumentTitle component

We cannot properly keep the previous title, as component are not
necessarily unmounted in the same order they were mounted.

Instead of trying to be smart about this, we now reset the page title
when `DocumentTitle` is unmounted, and consumers of this component can
mount another instance with a different title if they want to customize
the title. We can make this a bit smarted in the future if we need to,
but right now is good enough.

* Only add getting started title once