Fix quickvalues and field statistics refresh
Before this change, resetting `forceFetch` to `false` in
`AppWithSearchBar` had no immediate effect, as it required a re-render
for the new value to be passed to all children. The effect was that we
were fetching data more times than needed.

This commit moves `forceFetch` to the state of `AppWithSearchBar`, so
once we reset the value to `false`, all children will see the new value
and stop refreshing their data.

Fixes #3086