IndexSet default setting (#3209)
* Create IndexSetsResource#setDefault() API endpoint
* Remove default flag from IndexSetConfig in favor of a cluster config setting

This removes the `default` field from the IndexSetConfig database object
and creates a DefaultIndexSetConfig object in the cluster config
instead. Using a cluster config object makes it easier to ensure that
only one index set is the default.

There is no migration to convert the `default` flag to a
DefaultIndexSetConfig in the cluster config because this only happened
to people using alphas and betas and the user is able to just set the
default index set via the API or in the UI. (UI upcoming)

Also implement the IndexSetsResource#setDefault() method to actually set
the default in cluster config.

* Add UI elements to set a default index set
* Migrate "default" field to DefaultIndexSetConfig cluster config
* Change IndexSet registry and service interfaces for #getDefault() method

Since the default index must be there, we do not return an Optional but
throw an IllegalStateException if the default index set cannot be found.