Remove pluggable alert condition forms
Having custom pluggable forms, and falling back to using
`ConfigurationForm` when none was provided sounded like a great idea,
but in the end it is tricky, as forms behaved in slightly different
ways, and they also asume the developer would follow certain implied
conventions.

As this way also feels inconsistent from a user's point of view, I
think it is best at this point to only use `ConfigurationForm` components
for now, and think of a better way of doing custom forms for alert
conditions in future versions.