Micro batch loading grace periods
In #2972 we made the stream page loading much faster, but it in our
setup it is still slow (3-10 seconds). After some benchmarking I figured
that most of the time is spent in the "in grace period" check.

I tried to see if that field is anywhere used, but I found not usage for
it. Nonetheless removing it from the API seems a little bit too much for
me.

Therefore I started to implement a "micro" batching for this. So, in the
first step load all "last triggered_at" datetime for all given
conditions and streams and than calculate the "in grace" from that.