Showing index set in stream listing only if user is permitted to. (#3300)
This is fixing an issue where it's implicitly assumed that a user
navigating to the streams page has the required permissions to list all
index sets. This might not be the case for non-admin users, leading to
the streams page breaking completely.

This commit adds a check for the permission and the presence of the index
set definition in the component showing the stream summary, rendering
index set details only if both are present.

Fixes #3299.