Refreshing saved searches store after deleting one.
Before this change, the saved searches store was not automatically
refreshed after triggering a successfull deletion. This lead to the
saved searches dropdown in the search bar being inconsistent.

This change triggers a refresh after a successfull deletion, so the
store and therefore the saved searches dropdown is accurate.

Fixes #2841.