Ignore reopened indices for count-based retention strategies.
Before this change, reopening a closed index while a count-based
retention strategy is configured leads to:

a) a possible higher server log volume, as the index count is probably
higher than the configured limit, but a reopened index is ignored during
retention

b) an incorrect number of indices being deleted, as reopened indices
are counted against the maximum number of indices, but ignored during
retention

With this change, reopened indices are being filtered out when
determining the number of current indices which is compared against the
maxmimum number of indices _and_ when building the list of indices to
delete during retention.

Fixes #2405, #3297.